feature: change code to reflect config changes
