+++ mdbackup/config/__init__.py
@@ -1,7 +1,11 @@
-from mdbackup.config.configuration import Config, SecretConfig, StorageConfig
+from mdbackup.config.cloud import CloudConfig
+from mdbackup.config.configuration import Config
+from mdbackup.config.secret import SecretConfig
+from mdbackup.config.storage import StorageConfig
 
 __all__ = [
     Config,
+    CloudConfig,
     SecretConfig,
     StorageConfig,
 ]
